support-dialogue-gen/requirements.txt
pydantic>=2.0
requests
python-dotenv
tqdm
uuid
nlpaug
textattack
datasets
